# Improve chain interceptors performance by ~20%