Improve chain interceptors performance by ~20%

This is done in 3 ways, each contributing to the improvements above:

- Use a sync.Pool to avoid allocating the state struct on every request
- Add to the state struct all variables that would be otherwise captured
  by the state.next closure
- Optimize the code of the state.next function a bit

It could be possible to bring the number of allocations down to 0, but
it would likely require the code to become much harder to read.

```
name                         old time/op    new time/op    delta
ChainUnaryInterceptor/1-4      3.34ns ± 9%    3.32ns ± 3%     ~     (p=0.841 n=5+5)
ChainUnaryInterceptor/3-4       147ns ± 3%     117ns ± 3%  -20.38%  (p=0.008 n=5+5)
ChainUnaryInterceptor/5-4       157ns ± 2%     129ns ± 3%  -18.07%  (p=0.008 n=5+5)
ChainUnaryInterceptor/10-4      183ns ± 6%     150ns ± 1%  -18.02%  (p=0.008 n=5+5)
ChainStreamInterceptor/1-4     3.05ns ± 5%    3.08ns ± 4%     ~     (p=0.548 n=5+5)
ChainStreamInterceptor/3-4     86.9ns ± 3%    65.4ns ± 6%  -24.66%  (p=0.008 n=5+5)
ChainStreamInterceptor/5-4     98.7ns ± 7%    74.2ns ± 2%  -24.82%  (p=0.008 n=5+5)
ChainStreamInterceptor/10-4     124ns ± 8%     100ns ± 4%  -18.83%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
ChainUnaryInterceptor/1-4       0.00B          0.00B          ~     (all equal)
ChainUnaryInterceptor/3-4       80.0B ± 0%     16.0B ± 0%  -80.00%  (p=0.008 n=5+5)
ChainUnaryInterceptor/5-4       80.0B ± 0%     16.0B ± 0%  -80.00%  (p=0.008 n=5+5)
ChainUnaryInterceptor/10-4      80.0B ± 0%     16.0B ± 0%  -80.00%  (p=0.008 n=5+5)
ChainStreamInterceptor/1-4      0.00B          0.00B          ~     (all equal)
ChainStreamInterceptor/3-4      80.0B ± 0%     16.0B ± 0%  -80.00%  (p=0.008 n=5+5)
ChainStreamInterceptor/5-4      80.0B ± 0%     16.0B ± 0%  -80.00%  (p=0.008 n=5+5)
ChainStreamInterceptor/10-4     80.0B ± 0%     16.0B ± 0%  -80.00%  (p=0.008 n=5+5)

name                         old allocs/op  new allocs/op  delta
ChainUnaryInterceptor/1-4        0.00           0.00          ~     (all equal)
ChainUnaryInterceptor/3-4        2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
ChainUnaryInterceptor/5-4        2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
ChainUnaryInterceptor/10-4       2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
ChainStreamInterceptor/1-4       0.00           0.00          ~     (all equal)
ChainStreamInterceptor/3-4       2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
ChainStreamInterceptor/5-4       2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
ChainStreamInterceptor/10-4      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
```

RELEASE NOTES: None