@dfawley 2021-12-07
> it was such an obvious self-contained and low-hanging fruit that it seemed worthwhile.

Unfortunately it comes at the cost of making the code less readable, so it might not be a worthwhile tradeoff depending on the numbers.  We probably already went too far with the previous set of optimizations IMO.

> Will try to run these after adding the chains. Assuming they show anything worthwhile, do you want me to include the changes to the benchmark harness in this PR?

Yes, for sure.  Unless the conclusion is "optimizing these will never be worth doing", then it would be a good idea to include them even if this isn't the right approach.  Also, I'm totally fine with choosing the benchmark parameters that highlight the improvement here.  (E.g. set request/response payload size to 1 byte and use 10ish instances of a dummy interceptor to have a long chain.)